streamlined XCM Manager driven analytics in xcmmatch
* indexer inserts xcmtransfer/xcmmessages/xcmtransferdestcandidate for the tip and any old hour
* but all *BULK* XCM analytics in XCMManager (none in Crawler/Indexer) with single "xcmmatch" service
* reduced some SQL load by only running bulk analytics when new messages appear
* XCM v0+v1+v2 coverage with single instruction module
