+++ substrate/indexer.js
@@ -1202,321 +1202,6 @@ module.exports = class Indexer extends AssetManager {
         }
     }
 
-    /*
-        updateXCMTransferDestCandidate(bn, blockTS, extrinsicID, eventIndex, pallet, method, fromAddress, asset, amountReceived, paraIDs, caller = false) {
-            let eventID = `${this.chainID}-${extrinsicID}-${eventIndex}`;
-            this.xcmtransferdestcandidate[eventID] = {
-                chainIDDest: this.chainID,
-                eventID,
-                pallet,
-                method,
-                fromAddress,
-                blockNumberDest: bn,
-                asset,
-                destTS: blockTS,
-                amountReceived,
-                paraIDs: JSON.stringify(Object.keys(paraIDs))
-            }
-            if (caller) {
-                console.log(`${caller} candidate`, this.xcmtransferdestcandidate[eventID]);
-            }
-        }
-    */
-
-    /*
-    xcmMatch matches cross transfers between SENDING events held in "xcmtransfer"
-        mysql> desc xcmtransfer;
-        +-----------------+---------------------------------------------------------------------------------+------+-----+----------------------+-------+
-        | Field           | Type                                                                            | Null | Key | Default              | Extra |
-        +-----------------+---------------------------------------------------------------------------------+------+-----+----------------------+-------+
-        | extrinsicHash   | varchar(67)                                                                     | NO   | PRI | NULL                 |       |
-        | extrinsicID     | varchar(32)                                                                     | NO   |     | NULL                 |       |
-        | chainID         | int(11)                                                                         | NO   |     | NULL                 |       |
-        | chainIDDest     | int(11)                                                                         | NO   |     | NULL                 |       |
-        | blockNumber     | varchar(67)                                                                     | NO   |     | NULL                 |       |
-        | fromAddress     | varchar(67)                                                                     | YES  |     | NULL                 |       |
-        | asset           | varchar(80)                                                                     | YES  |     | NULL                 |       |
-        | blockNumberDest | int(11)                                                                         | YES  |     | NULL                 |       |
-        | sourceTS        | int(11)                                                                         | YES  |     | NULL                 |       |
-        | destTS          | int(11)                                                                         | YES  |     | NULL                 |       |
-        | amountSent      | decimal(36,18)                                                                  | YES  |     | NULL                 |       |
-        | amountReceived  | decimal(36,18)                                                                  | YES  |     | 0.000000000000000000 |       |
-        | status          | enum('NonFinalizedSource','FinalizedSource','NonFinalizedDest','FinalizedDest') | YES  |     | NonFinalizedSource   |       |
-        +-----------------+---------------------------------------------------------------------------------+------+-----+----------------------+-------+
-        13 rows in set (0.00 sec)
-
-    and CANDIDATE destination events (from various xcm received messages on a destination chain) that are stored here:
-
-        mysql> desc xcmtransferdestcandidate;
-        +-----------------+----------------+------+-----+---------+-------+
-        | Field           | Type           | Null | Key | Default | Extra |
-        +-----------------+----------------+------+-----+---------+-------+
-        | eventID         | varchar(67)    | NO   | PRI | NULL    |       |
-        | chainIDDest     | int(11)        | NO   |     | NULL    |       |
-        | fromAddress     | varchar(67)    | YES  |     | NULL    |       |
-        | blockNumberDest | int(11)        | YES  |     | NULL    |       |
-        | asset           | varchar(80)    | YES  |     | NULL    |       |
-        | destTS          | int(11)        | YES  |     | NULL    |       |
-        | amountReceived  | decimal(36,18) | YES  |     | NULL    |       |
-        | paraIDs         | blob           | YES  |     | NULL    |       |
-        +-----------------+----------------+------+-----+---------+-------+
-    */
-
-    async xcmAUSDUpdate() {
-        let kUSDTransferSql = `update xcmtransfer set nativeAssetChain = '{"Token":"KUSD"}~22000' where (asset like '{"Token":"%aUSD%"}' or asset like '{"Token":"%kUSD%"}') and asset not like "%-%" and relaychain = 'kusama' and nativeAssetChain is null;`
-        let kUSDCandidateSql = `update xcmtransferdestcandidate set nativeAssetChain = '{"Token":"KUSD"}~22000' where (asset like '{"Token":"%aUSD%"}' or asset like '{"Token":"%kUSD%"}') and asset not like "%-%" and (chainIDDest = 2 or chainIDDest > 20000) and nativeAssetChain is null;`
-        let aUSDTransferSql = `update xcmtransfer set nativeAssetChain = '{"Token":"AUSD"}~2000' where (asset like '{"Token":"%aUSD%"}' or asset like '{"Token":"%kUSD%"}') and asset not like "%-%" and relaychain = 'polkadot' and nativeAssetChain is null;`
-        let aUSDCandidateSql = `update xcmtransferdestcandidate set nativeAssetChain = '{"Token":"AUSD"}~2000' where (asset like '{"Token":"%aUSD%"}' or asset like '{"Token":"%kUSD%"}') and asset not like "%-%" and (chainIDDest !=2 and chainIDDest < 10000) and nativeAssetChain is null;`
-        this.batchedSQL.push(kUSDTransferSql);
-        this.batchedSQL.push(kUSDCandidateSql);
-        this.batchedSQL.push(aUSDTransferSql);
-        this.batchedSQL.push(aUSDCandidateSql);
-        await this.update_batchedSQL();
-    }
-
-    async xcmMatch(startTS = 1626145851, endTS = false, ratMin = .99) {
-        if (!endTS) endTS = this.currentTS();
-        for (let ts = startTS; ts < endTS; ts += 86400) {
-            await this.write_feedxcmdest(ts, ts + 86400, ratMin)
-        }
-    }
-
-    // This is the key workhorse that matches events on the destination chain (stored in xcmtransferdestcandidate) in some time interval
-    async write_feedxcmdest(startTS, endTS, ratMin = .99, lookbackSeconds = 7200) {
-        // match xcmtransferdestcandidate of the last 2 hours
-        //   (a) > 95% amountReceived / amountSent
-        //   (b) asset match
-        //   (c) time difference matching has to be less than 7200 (and greater than 0)
-        //   (d) TODO: require xcmtransferdestcandidate.paraIDs to match xcmtransfer.chainIDDest (this is NOT guarateed to be present)
-        // In case of ties, the FIRST one ( "order by diffTS" ) covers this
-        let sql = `select
-          chainID, extrinsicHash, d.chainIDDest, d.fromAddress, xcmtransfer.asset, d.paraIDs,
-          (d.destts - xcmtransfer.sourceTS) as diffTS,
-          xcmtransfer.extrinsicID,
-          xcmtransfer.amountSent,
-          xcmtransfer.transferIndex,
-          xcmtransfer.sourceTS,
-          xcmtransfer.fromAddress as senderAddress,
-          xcmtransfer.destAddress,
-          xcmtransfer.msgHash,
-          d.eventID,
-          d.extrinsicID as destExtrinsicID,
-          d.blockNumberDest,
-          d.amountReceived,
-          d.destTS,
-          d.amountReceived / xcmtransfer.amountSent as rat
-        from xcmtransfer, xcmtransferdestcandidate as d
- where  d.fromAddress = xcmtransfer.destAddress and
-        d.chainIDDest = xcmtransfer.chainIDDest and
-        ((d.asset = xcmtransfer.asset) or (d.nativeAssetChain = xcmtransfer.nativeAssetChain and d.nativeAssetChain is not null)) and
-        xcmtransfer.sourceTS >= ${startTS} and
-        xcmtransfer.sourceTS < ${endTS} and
-        d.destTS >= ${startTS} and
-        d.destTS < ${endTS+lookbackSeconds} and
-        xcmtransfer.matched = 0 and
-        d.matched = 0 and
-        xcmtransfer.incomplete = 0 and
-        d.destTS - xcmtransfer.sourceTS >= 0 and
-        d.destTS - xcmtransfer.sourceTS < ${lookbackSeconds} and
-        length(xcmtransfer.extrinsicID) > 0 and
-        xcmtransfer.amountSent >= d.amountReceived
-having ( ( rat > ${ratMin} and rat <= 1.0 ) or
-(asset = '{"Token":"DOT"}' and amountSent - amountReceived < 500000000) or
-(asset = '{"Token":"KSM"}' and amountSent - amountReceived < 1000000000) or
-(asset = '{"Token":"KAR"}' and amountSent - amountReceived < 10000000000 ) )
-order by chainID, extrinsicHash, diffTS`
-        //console.log("match_xcm", sql)
-        try {
-            let xcmmatches = await this.poolREADONLY.query(sql);
-            let addressextrinsic = []
-            let hashes = []
-            let matched = {}
-            let matches = 0;
-            if (xcmmatches.length > 0) {
-                console.log("[Found] match_xcm", sql)
-            } else {
-                //enable this for debugging
-                //console.log("[Empty] match_xcm", sql)
-            }
-            for (let i = 0; i < xcmmatches.length; i++) {
-                let d = xcmmatches[i];
-                if (matched[d.extrinsicHash] == undefined && matched[d.eventID] == undefined) {
-                    let priceUSD = 0;
-                    let amountSentUSD = 0;
-                    let amountReceivedUSD = 0;
-                    let decimals = this.getAssetDecimal(d.asset, d.chainID)
-                    if (decimals === false) {
-                        decimals = this.getAssetDecimal(d.asset, d.chainIDDest)
-                    }
-                    if (decimals !== false) {
-                        let [_, __, priceUSDsourceTS] = await this.computeUSD(1.0, d.asset, d.chainID, d.sourceTS);
-                        if (priceUSDsourceTS > 0) {
-                            priceUSD = priceUSDsourceTS;
-                            let amountSent = parseFloat(d.amountSent) / 10 ** decimals;
-                            let amountReceived = parseFloat(d.amountReceived) / 10 ** decimals;
-                            amountSentUSD = (amountSent > 0) ? priceUSD * amountSent : 0;
-                            amountReceivedUSD = (amountReceived > 0) ? priceUSD * amountReceived : 0;
-                        }
-                    }
-                    let sql = `update xcmtransfer
-            set blockNumberDest = ${d.blockNumberDest},
-                destTS = ${d.destTS},
-                amountSentUSD = '${amountSentUSD}',
-                amountReceived = '${d.amountReceived}',
-                amountReceivedUSD = '${amountReceivedUSD}',
-                priceUSD = '${priceUSD}',
-                matched = 1,
-                matchedExtrinsicID = '${d.destExtrinsicID}',
-                matchedEventID = '${d.eventID}'
-            where extrinsicHash = '${d.extrinsicHash}' and transferIndex = '${d.transferIndex}'`
-                    this.batchedSQL.push(sql);
-                    matches++;
-                    // (a) with extrinsicID, we get both the fee (rat << 1) ANDthe transferred item for when one is isFeeItem=0 and another is isFeeItem=1; ... otherwise we get (b) with just the eventID
-                    let w = d.destExtrinsicID && (d.destExtrinsicID.length > 0) ? `extrinsicID = '${d.destExtrinsicID}'` : `eventID = '${d.eventID}'`;
-                    let sql2 = `update xcmtransferdestcandidate set matched = 1, matchedExtrinsicHash = '${d.extrinsicHash}' where ${w}`
-                    console.log(sql2);
-                    this.batchedSQL.push(sql2);
-                    // never match more than once, by marking both the sending record and the receiving record
-                    matched[d.extrinsicHash] = true;
-                    matched[d.eventID] = true;
-
-                    // problem: this does not store the "dust" fee (which has rat << 1) in bigtable
-                    let match = {
-                        chainID: d.chainID,
-                        chainIDDest: d.chainIDDest,
-                        blockNumberDest: d.blockNumberDest,
-                        asset: d.asset,
-                        amountSent: d.amountSent,
-                        amountReceived: d.amountReceived,
-                        fromAddress: d.senderAddress, //from xcmtransfer.fromAddress
-                        destAddress: d.fromAddress, //from xcmtransferdestcandidate.fromAddress
-                        msgHash: (d.msgHash != undefined) ? d.msgHash : '0x',
-                        extrinsicHash: d.extrinsicHash,
-                        extrinsicID: d.extrinsicID,
-                        destExtrinsicID: d.destExtrinsicID,
-                        eventID: d.eventID,
-                        sourceTS: d.sourceTS,
-                        destTS: d.destTS
-                    }
-
-                    // 1. write "addressExtrinsic" feedxcmdest with row key address-extrinsicHash and column extrinsicHash#chainID-extrinsicID-transferIndex
-                    //    * we use the SENDING chainID + the sender extrinsicID as the column for the "feedxcmdest" column family
-                    let rowKey = `${d.fromAddress.toLowerCase()}#${d.extrinsicHash}` // NOTE: this matches "feed"
-                    let cres = {
-                        key: rowKey,
-                        data: {
-                            feedxcmdest: {},
-                        }
-                    };
-                    let extrinsicHashEventID = `${d.extrinsicHash}#${d.chainID}-${d.extrinsicID}-${d.transferIndex}`
-                    cres['data']['feedxcmdest'][extrinsicHashEventID] = {
-                        value: JSON.stringify(match),
-                        timestamp: d.sourceTS * 1000000 // NOTE: to support rematching, we do NOT use d.destTS
-                    };
-                    addressextrinsic.push(cres);
-
-                    // 2. write "hashes" feedxcmdest with row key extrinsicHash and column extrinsicHash#chainID-extrinsicID (same as 1)
-                    let hres = {
-                        key: d.extrinsicHash,
-                        data: {
-                            feedxcmdest: {}
-                        }
-                    }
-                    hres['data']['feedxcmdest'][extrinsicHashEventID] = {
-                        value: JSON.stringify(match),
-                        timestamp: d.sourceTS * 1000000 // NOTE: to support rematching, we do NOT use d.destTS
-                    };
-                    console.log("MATCH#", matches, "hashes rowkey", d.extrinsicHash, "col", "addressExtrinsic rowkey=", rowKey, "col", extrinsicHashEventID);
-                    hashes.push(hres)
-                } else {
-                    //console.log("SKIP", matched[d.extrinsicHash], matched[d.eventID]);
-                }
-            }
-            if (addressextrinsic.length > 0) {
-                await this.insertBTRows(this.btAddressExtrinsic, addressextrinsic, "addressextrinsic")
-            }
-            if (hashes.length > 0) {
-                await this.insertBTRows(this.btHashes, hashes, "addressextrinsic")
-            }
-            if (addressextrinsic.length > 0 || hashes.length > 0) {
-                await this.update_batchedSQL();
-            }
-        } catch (err) {
-            console.log("WRITE_FEEDXCMDEST", err)
-        }
-        let logDT = new Date(startTS * 1000)
-        console.log(`match_xcm ${startTS} covered ${logDT}`)
-    }
-
-    // This is the key workhorse that matches xcmmessages with
-    // match xcmmessages incoming = 0 with incoming = 1
-    //   (a) msgHash + chainID + chainIDDest matching
-    //   (b) time difference between blockTS matching has to be less than 120 (lookbackSeconds parameter)
-    //   (c)
-    // In case of ties, the FIRST one ( "order by diffTS" ) covers this
-    async xcmmessages_match(startTS, endTS, lookbackSeconds = 120) {
-        let sql = `select
-          s.msgHash, s.blockNumber as s_blockNumber, d.blockNumber as d_blockNumber, s.sentAt as s_sentAt, d.sentAt as d_sentAt, s.chainID, s.chainIDDest, d.blockTS as destTS, s.blockTS as sourceTS, abs(d.blockTS - s.blockTS) as diffTS, (d.sentAt - s.sentAt) as diffSentAt
-        from xcmmessages as s, xcmmessages as d
- where  d.msgHash = s.msgHash and
-        d.chainID = s.chainID and
-        d.chainIDDest = s.chainIDDest and
-        s.incoming = 0 and
-        d.incoming = 1 and
-        s.blockTS >= ${startTS} and
-        s.blockTS < ${endTS} and
-        d.blockTS >= ${startTS} and
-        d.blockTS < ${endTS+lookbackSeconds} and
-        s.matched = 0 and
-        d.matched = 0
-having (diffSentAt >= 0 and diffSentAt <= 4)
-order by msgHash, diffSentAt, diffTS`
-        //console.log("xcmmessages_match", sql)
-        try {
-            let xcmmatches = await this.poolREADONLY.query(sql);
-            let matched = {}
-            if (xcmmatches.length > 0) {
-                console.log(`[Found] xcmmessages_match ${xcmmatches.length}`, sql)
-            } else {
-                //enable this for debugging
-                //console.log("[Empty] match_xcm", sql)
-            }
-            let vals = ["sourceTS", "destTS", "matched", "sourceSentAt", "destSentAt", "sourceBlocknumber", "destBlocknumber", "matchDT"];
-            let out = [];
-            for (let i = 0; i < xcmmatches.length; i++) {
-                let s = xcmmatches[i];
-                // to protect against the same dest message matched more than once we keep in the "matched" map a set of msgHash-sentAt (for xcmmessages dest candidates)
-                // Note in case of multiple matches, the "order by diffTS" in the SQL statment picks the FIRST one in time closest with the smallest diffTS
-                let k = `${s.msgHash}:${s.d_sentAt}`
-                if (matched[k] == undefined) {
-                    out.push(`('${s.msgHash}', ${s.s_blockNumber}, 0, ${s.sourceTS}, ${s.destTS}, 1, '${s.s_sentAt}', '${s.d_sentAt}', '${s.s_blockNumber}', '${s.d_blockNumber}', Now())`)
-                    out.push(`('${s.msgHash}', ${s.d_blockNumber}, 1, ${s.sourceTS}, ${s.destTS}, 1, '${s.s_sentAt}', '${s.d_sentAt}', '${s.s_blockNumber}', '${s.d_blockNumber}', Now())`)
-                    matched[k] = true;
-                } else {
-                    // console.log("NO MATCH", s);
-                }
-            }
-            let logDT = new Date(startTS * 1000)
-            console.log(`xcmmessages_match ${startTS} covered ${logDT} MATCHES: `, out.length);
-            await this.upsertSQL({
-                "table": "xcmmessages",
-                "keys": ["msgHash", "blockNumber", "incoming"],
-                "vals": vals,
-                "data": out,
-                "replace": vals
-            });
-        } catch (err) {
-            console.log("xcmmessages_match", err)
-        }
-    }
-
-    async xcmmessagesMatch(startTS = 1626145851, endTS = false, lookbackSeconds = 120) {
-        if (!endTS) endTS = this.currentTS();
-        for (let ts = startTS; ts < endTS; ts += 86400) {
-            await this.xcmmessages_match(ts, ts + 86400, lookbackSeconds)
-        }
-    }
-
     knownParaID(paraID) {
         return this.paraIDs.includes(paraID);
     }
