--- substrate/xcmanalytics
@@ -1,24 +1,0 @@
-#!/usr/bin/env node
- // Usage:  xcmanalytics
-var XCMManager = require("./xcmmanager");
-
-async function main() {
-    var manager = new XCMManager();
-    await manager.xcm_init()
-    let lookbackDays = 3;
-    process.argv.forEach(function(val, index, array) {
-        if (index == 2 && val.length > 0) {
-            lookbackDays = parseInt(val, 10);
-        }
-    });
-    await manager.xcmanalytics(lookbackDays)
-}
-
-main()
-    .then(() => {
-        process.exit(0);
-    })
-    .catch((e) => {
-        console.error('ERROR', e);
-        process.exit(1);
-    });
