+++ views/chain.ejs
@@ -80,9 +80,8 @@ for (let i=0; i<tabs.length; i++) {
 	  <th>Msg Type</th>
 	  <th>Sent</th>
 	  <th>Received</th>
-	  <th>Instructions</th>
+          <th>Assets Received/Instructions</th>
           <th>Beneficiary</th>
-          <th>Assets Received</th>
 	  <th>Matched/Time</th>
         </tr>
       </thead>
